SandwichClub/requirements.txt
Django==1.10.5
django-bootstrap-toolkit==2.15.0
django-registration-redux==1.4
olefile==0.44
Pillow==4.0.0
